backend/pyproject.toml
[tool.poetry]
name = "backend-v2"
version = "0.1.0"
description = ""
authors = ["Samrat Dutta Roy <samrat.roy@vinsol.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
django-cors-headers = "^4.6.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
django-eventstream = "^5.3.1"
daphne = "^4.1.2"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}
graphene-django = "^3.2.2"


[tool.poetry.group.test.dependencies]
pylint-django = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
